# Apply alpha in nonlinear space for images and async video