### Description
Apply alpha in nonlinear space for compatibility with bad applications.

The image source now needs to be a custom draw, which prevents the direct filter rendering optimization.

We may want to provide a toggle in the future, but just try to match previous behavior for now. v27 has enough new features for support to deal with.

### Motivation and Context
Alpha should be applied in linear space, but most applications do not, including Chromium and Visual Studio.

### How Has This Been Tested?
Used 255/255/255/64 texture in various scenarios.

The shader involved in more complicated, but the performance hit seems manageable:
NVIDIA RTX 2080 Ti: 34 us -> 41 us for 1080p image
Intel UHD Graphics 750: 448 us -> 576 us for 1080p image

### Types of changes
- Bug fix (non-breaking change which fixes an issue)

### Checklist:
- [x] My code has been run through [clang-format](https://github.com/obsproject/obs-studio/blob/master/.clang-format).
- [x] I have read the [**contributing** document](https://github.com/obsproject/obs-studio/blob/master/CONTRIBUTING.rst).
- [x] My code is not on the master branch.
- [x] The code has been tested.
- [x] All commit messages are properly formatted and commits squashed where appropriate.
- [x] I have included updates to all appropriate documentation.